Simple Go app for Consul-Redis integration

+++ consul_dev_cluster2/scripts/consul.sh
@@ -18,11 +18,26 @@ which consul &>/dev/null || {
     if [ -z "$CONSUL_DEMO_VERSION" ]; then
         CONSUL_DEMO_VERSION=$(curl -s "${CHECKPOINT_URL}"/consul | jq .current_version | tr -d '"')
     fi
-
-    echo "Fetching Consul version ${CONSUL_DEMO_VERSION} ..."
-    pushd /tmp/
-    curl -s https://releases.hashicorp.com/consul/${CONSUL_DEMO_VERSION}/consul_${CONSUL_DEMO_VERSION}_linux_amd64.zip -o consul.zip
-
+	
+	pushd /tmp/
+	
+    if [ -f "/vagrant/pkg/consul_${CONSUL_DEMO_VERSION}_linux_amd64.zip" ]; then
+		echo "Found Consul in /vagrant/pkg"
+		cp /vagrant/pkg/consul_${CONSUL_DEMO_VERSION}_linux_amd64.zip /tmp/consul.zip
+    else
+		echo "Fetching Consul version ${CONSUL_DEMO_VERSION} ..."
+		
+		curl -s https://releases.hashicorp.com/consul/${CONSUL_DEMO_VERSION}/consul_${CONSUL_DEMO_VERSION}_linux_amd64.zip -o consul.zip
+		
+		if [ $? -ne 0 ]; then
+			echo "Download failed! Exiting."
+			exit 1
+		fi
+		
+		# Copying the archive in the /vagrant folder to reuse it for future provisionings or other VMs
+		sudo cp consul.zip /vagrant/pkg/consul_${CONSUL_DEMO_VERSION}_linux_amd64.zip
+   fi
+    
     echo "Installing Consul version ${CONSUL_DEMO_VERSION} ..."
     unzip consul.zip
     sudo chmod +x consul
